+++ TwidereSDK/Sources/TwidereUI/Content/PollOptionView+ViewModel.swift
@@ -253,112 +253,3 @@ extension PollOptionView.ViewModel {
             .store(in: &disposeBag)
     }
 }
-
-extension PollOptionView {
-    public typealias ConfigurationContext = StatusView.ConfigurationContext
-}
-
-extension PollOptionView {
-    public func configure(
-        pollOption option: MastodonPollOption,
-        configurationContext: ConfigurationContext
-    ) {
-        viewModel.objects.insert(option)
-        
-        configurationContext.authenticationContext
-            .assign(to: \.authenticationContext, on: viewModel)
-            .store(in: &disposeBag)
-
-        // metaContent
-        Publishers.CombineLatest(
-            option.poll.status.publisher(for: \.emojis),
-            option.publisher(for: \.title)
-        )
-        .map { emojis, title -> MetaContent? in
-            do {
-                let content = MastodonContent(content: title, emojis: emojis.asDictionary)
-                let metaContent = try MastodonMetaContent.convert(document: content)
-                return metaContent
-            } catch {
-                assertionFailure()
-                return PlaintextMetaContent(string: title)
-            }
-        }
-        .assign(to: \.metaContent, on: viewModel)
-        .store(in: &disposeBag)
-
-        // $isExpire
-        option.poll.publisher(for: \.expired)
-            .assign(to: \.isExpire, on: viewModel)
-            .store(in: &disposeBag)
-        // isMultiple
-        viewModel.isMultiple = option.poll.multiple
-        
-        let optionIndex = option.index
-        let authorDomain = option.poll.status.author.domain
-        let authorUserID = option.poll.status.author.id
-        // isSelect, isPollVoted, isMyPoll
-        Publishers.CombineLatest4(
-            option.publisher(for: \.poll),
-            option.publisher(for: \.voteBy),
-            option.publisher(for: \.isSelected),
-            viewModel.$authenticationContext
-        )
-        .sink { [weak self] poll, optionVoteBy, isSelected, authenticationContext in
-            guard let self = self else { return }
-            
-            let domain: String
-            let userID: String
-            switch authenticationContext {
-            case .twitter, .none:
-                domain = ""
-                userID = ""
-            case .mastodon(let authenticationContext):
-                domain = authenticationContext.domain
-                userID = authenticationContext.userID
-            }
-
-            let options = poll.options
-            let pollVoteBy = poll.voteBy
-
-            let isMyPoll = authorDomain == domain
-                        && authorUserID == userID
-
-            let votedOptions = options.filter { option in
-                option.voteBy.contains(where: { $0.id == userID && $0.domain == domain })
-            }
-            let isRemoteVotedOption = votedOptions.contains(where: { $0.index == optionIndex })
-            let isRemoteVotedPoll = pollVoteBy.contains(where: { $0.id == userID && $0.domain == domain })
-
-            let isLocalVotedOption = isSelected
-
-            let isSelect: Bool? = {
-                if isLocalVotedOption {
-                    return true
-                } else if !votedOptions.isEmpty {
-                    return isRemoteVotedOption ? true : false
-                } else if isRemoteVotedPoll, votedOptions.isEmpty {
-                    // the poll voted. But server not mark voted options
-                    return nil
-                } else {
-                    return false
-                }
-            }()
-            self.viewModel.isSelect = isSelect
-            self.viewModel.isPollVoted = isRemoteVotedPoll
-            self.viewModel.isMyPoll = isMyPoll
-        }
-        .store(in: &disposeBag)
-        // percentage
-        Publishers.CombineLatest(
-            option.poll.publisher(for: \.votesCount),
-            option.publisher(for: \.votesCount)
-        )
-        .map { pollVotesCount, optionVotesCount -> Double? in
-            guard pollVotesCount > 0, optionVotesCount >= 0 else { return 0 }
-            return Double(optionVotesCount) / Double(pollVotesCount)
-        }
-        .assign(to: \.percentage, on: viewModel)
-        .store(in: &disposeBag)
-    }
-}
